create a seperate make_app.py for app.py to make the structure more concise
